new component board + new background image
